Only create AzureMonitorClient once
