+++ src/Promitor.Scraper.Host/Scheduling/MetricScrapingJob.cs
@@ -3,13 +3,16 @@ using System.Threading;
 using System.Threading.Tasks;
 using CronScheduler.Extensions.Scheduler;
 using GuardNet;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using Promitor.Core;
 using Promitor.Core.Scraping.Configuration.Model;
 using Promitor.Core.Scraping.Configuration.Model.Metrics;
 using Promitor.Core.Scraping.Configuration.Providers.Interfaces;
 using Promitor.Core.Scraping.Factories;
 using Promitor.Core.Scraping.Prometheus.Interfaces;
 using Promitor.Core.Telemetry.Metrics.Interfaces;
+using Promitor.Integrations.AzureMonitor;
 
 namespace Promitor.Scraper.Host.Scheduling
 {
@@ -19,6 +22,8 @@ namespace Promitor.Scraper.Host.Scheduling
         private readonly IMetricsDeclarationProvider _metricsDeclarationProvider;
         private readonly IPrometheusMetricWriter _prometheusMetricWriter;
         private readonly IRuntimeMetricsCollector _runtimeMetricsCollector;
+        private readonly AzureMonitorClient _azureMonitorClient;
+        private readonly IConfiguration _configuration;
         private readonly ILogger _logger;
 
         private readonly MetricScraperFactory _metricScraperFactory;
@@ -28,6 +33,7 @@ namespace Promitor.Scraper.Host.Scheduling
             IPrometheusMetricWriter prometheusMetricWriter,
             IRuntimeMetricsCollector runtimeMetricsCollector,
             MetricScraperFactory metricScraperFactory,
+            IConfiguration configuration,
             ILogger<MetricScrapingJob> logger)
         {
             Guard.NotNull(metric, nameof(metric));
@@ -41,14 +47,16 @@ namespace Promitor.Scraper.Host.Scheduling
             _metricsDeclarationProvider = metricsDeclarationProvider;
             _prometheusMetricWriter = prometheusMetricWriter;
             _runtimeMetricsCollector = runtimeMetricsCollector;
+            _configuration = configuration;
             _logger = logger;
 
             _metricScraperFactory = metricScraperFactory;
-
+            _azureMonitorClient = ConfigureAzureMonitorClient();
             ConfigureJob();
         }
 
         public string CronSchedule { get; set; }
+
         // ReSharper disable once UnassignedGetOnlyAutoProperty
         public string CronTimeZone { get; }
         public bool RunImmediately { get; set; }
@@ -78,8 +86,28 @@ namespace Promitor.Scraper.Host.Scheduling
         {
             _logger.LogInformation("Scraping {MetricName} for resource type {ResourceType}", metricDefinitionDefinition.PrometheusMetricDefinition.Name, metricDefinitionDefinition.Resource.ResourceType);
 
-            var scraper = _metricScraperFactory.CreateScraper(metricDefinitionDefinition.Resource.ResourceType, azureMetadata, _prometheusMetricWriter, _runtimeMetricsCollector);
+            var scraper = _metricScraperFactory.CreateScraper(metricDefinitionDefinition.Resource.ResourceType, azureMetadata, _prometheusMetricWriter, _azureMonitorClient);
             await scraper.ScrapeAsync(metricDefinitionDefinition);
+        }
+
+        private AzureMonitorClient ConfigureAzureMonitorClient()
+        {
+            var azureCredentials = DetermineAzureCredentials();
+            var azureMetadata = _metricsDeclarationProvider.Get().AzureMetadata;
+            var azureMonitorClient = new AzureMonitorClient(azureMetadata.Cloud, azureMetadata.TenantId, azureMetadata.SubscriptionId, azureCredentials.ApplicationId, azureCredentials.Secret, _runtimeMetricsCollector, _logger);
+            return azureMonitorClient;
+        }
+
+        private AzureCredentials DetermineAzureCredentials()
+        {
+            var applicationId = _configuration.GetValue<string>(EnvironmentVariables.Authentication.ApplicationId);
+            var applicationKey = _configuration.GetValue<string>(EnvironmentVariables.Authentication.ApplicationKey);
+
+            return new AzureCredentials
+            {
+                ApplicationId = applicationId,
+                Secret = applicationKey
+            };
         }
     }
 }
